Fix prog.txt audit findings (points 1-5) (#256)
1. Dtype/precision: mps.py::_sample_bitstring, healing.py's
   MAX_SEMANTIC_DISTANCE, and observables.py::_apply_pauli_term_jax no
   longer bake in or hardcode a dtype independent of the ambient
   jax_enable_x64 state / the array they operate on.

2. Duplication: compiler.py's 1-qubit gate matrix table is now the one
   canonical _gate_1q_matrix, reused by kernels.py's _gate_matrix_elements
   and _build_distributed_chunk_step (was: 3 hand-copied switch tables).
   compiler.py's do_2q dispatch is now a single lax.switch instead of 5
   lax.cond calls + a jnp.where chain + an unconditionally-computed
   apply_swap. registry.py's dark_background side effect moved into an
   explicit apply_dark_theme(), no longer fired by a bare import.

3. API/doc consistency: entropy.py documents its nats (not log2/bits)
   convention explicitly; physics/__init__.py now exports
   decode_with_erasure_fallback and counts_in_intervals_dimension (were
   reachable from the top-level package but not the subpackage itself);
   qiskit_pennylane.py's _to_qiskit_bit_order is vectorized (loops over
   n_qubits, not 2**n_qubits); healing.py's Italian docstrings/comments
   translated. mitigation/__init__.py's _jit exports, cartesian.py's
   docstring, noise_model_from_qiskit_backend's approximation
   disclosure, and bounded_exponential_extrapolate's no-_jit comment
   were already correct -- no change needed. _js_divergence's p+eps
   formula is left as-is: its docstring documents an intentional exact
   match to an external validated reference.

4. Edge-case correctness: autodiff.py's energy_fn no longer collapses
   NoiseSpec(qubits=[]) (explicitly zero qubits) into qubits=None
   (apply to all) -- a real silent-wrong-answer bug, confirmed by a
   reproduction test before the fix. entropy.py::mutual_information now
   raises on overlapping qubits_a/qubits_b instead of silently
   corrupting the result. magic_entropy.py's _magic_entropy_core clips
   then renormalizes eigenvalues (matches magic_entropy_from_shadows'
   own pattern) instead of clipping alone. _engine_imports.py's
   deployment-fallback stub now raises explicit NotImplementedError
   instead of opaque AttributeError. noise_model_from_qiskit_backend's
   approximation and bridge.py::_apply_active_space (function no longer
   exists) needed no action.

5. Performance: kraus_channels.py::_qubit_index_pairs is now
   lru_cache'd (was rebuilt from scratch every shot in a Monte Carlo ZNE
   loop). qec.py::counts_in_intervals_dimension uses np.searchsorted
   (O(n log n)) instead of an O(n^2) per-reference-point brute-force
   scan. mps.py::get_top_k_probable_states uses heapq.nlargest instead
   of a full sort-then-slice. disk_overflow.py's ConditionalPhase now
   batches consecutive gates (one load/save per chunk for the whole
   run), matching LocalPhase's existing pattern, instead of one full
   disk cycle per individual gate.

Every fix has a dedicated regression test that failed before and passes
after. Full tests/unit suite: 897 passed, 0 failed, 21 skipped.

+++ dense_evolution/physics/observables.py
@@ -330,7 +330,7 @@ def _apply_pauli_term_jax(statevector, terms, inferred_n_qubits):
     indices = jnp.arange(dim)
     source_idx = indices ^ flip_mask
 
-    coeff = jnp.ones(dim, dtype=jnp.complex128)
+    coeff = jnp.ones(dim, dtype=statevector.dtype)
     for q, p in terms.items():
         bit = (source_idx >> bit_pos(q)) & 1
         if p == 'Y':
